+++ modulo4/filtros-ordenacao-paginacao/filtros-ordenacao-paginacao/src/data/GetTypes.ts
@@ -1,0 +1,9 @@
+import { connection } from "./connection"
+
+export const getTypes = async (type: string) => {
+    const resultado = await connection.raw(`
+      SELECT * FROM aula49_exercicio
+      where type LIKE "${type}"
+    `)
+    return resultado[0]
+}
