+++ modulo4/filtros-ordenacao-paginacao/filtros-ordenacao-paginacao/src/data/getNames.ts
@@ -1,0 +1,9 @@
+import { connection } from "./connection"
+
+export const getNames = async (name: string) => {
+    const resultado = await connection.raw(`
+      SELECT * FROM aula49_exercicio
+      where name LIKE "%${name}%" 
+    `)
+    return resultado[0]
+ }
